ApproveApi.API.Approve

Description

Operations

Approve

createPrompt

createPrompt

Arguments

:: (Consumes CreatePrompt MimeJSON, MimeRender MimeJSON CreatePromptRequest)

=> CreatePromptRequest
"createPromptRequest"

-> ApproveApiRequest CreatePrompt MimeJSON Prompt MimeJSON

POST /prompt

Sending a prompt

Creates a prompt and pushes it to the user (sends via email, sms, or other supported protocols).

AuthMethod: AuthBasicApiKey

data CreatePrompt

getPrompt

getPrompt

Arguments

:: Id
"id" - The identifier for a pending or completed prompt. This is returned when you create a prompt.

-> ApproveApiRequest GetPrompt MimeNoContent Prompt MimeJSON

GET /prompt/{id}

Retrieve a prompt

Retrieve the prompt object with the given ID.

AuthMethod: AuthBasicApiKey

data GetPrompt

getPromptStatus

getPromptStatus

Arguments

:: Id
"id" - The prompt identifier.

-> ApproveApiRequest GetPromptStatus MimeNoContent PromptStatus MimeJSON

GET /prompt/{id}/status

Check prompt status

Returns whether a prompt has been completed by the user. This request does not require authentication, and so can be used client-side without sharing API credentials.

data GetPromptStatus