ApproveApi.Model

Description

Parameter newtypes

Id

newtype Id

Constructors

Id

Fields

unId :: Text

LongPoll

newtype LongPoll

Constructors

LongPoll

Fields

unLongPoll :: Bool

Models

AnswerMetadata

data AnswerMetadata

AnswerMetadata

Constructors

AnswerMetadata

Fields

answerMetadataIpAddress :: !(Maybe Text)

"ip_address"

answerMetadataBrowser :: !(Maybe Text)

"browser"

answerMetadataOperatingSystem :: !(Maybe Text)

"operating_system"

mkAnswerMetadata :: AnswerMetadata

Construct a value of type AnswerMetadata (by applying it's required fields, if any)

CreatePromptRequest

data CreatePromptRequest

CreatePromptRequest

Constructors

CreatePromptRequest

Fields

createPromptRequestUser :: !Text

Required "user" - The user to send the approval request to. Can be either an email address or a phone number.

createPromptRequestBody :: !Text

Required "body" - The body of the approval request to show the user.

createPromptRequestTitle :: !(Maybe Text)

"title" - The title of an approval request. Defaults to an empty string.

createPromptRequestApproveText :: !(Maybe Text)

"approve_text" - The approve action text. Defaults to 'Approve'.

createPromptRequestApproveRedirectUrl :: !(Maybe Text)

"approve_redirect_url" - An HTTPS URL to redirect the user to if the prompt is approved. This URL is kept secret until the user is redirected to it.

createPromptRequestRejectText :: !(Maybe Text)

"reject_text" - The reject action text. If not specified the reject button will NOT be rendered, and the user will only see an approve action button.

createPromptRequestRejectRedirectUrl :: !(Maybe Text)

"reject_redirect_url" - An HTTPS URL to redirect the user to if the prompt is rejected. This URL is kept secret until the user is redirected to it.

createPromptRequestLongPoll :: !(Maybe Bool)

"long_poll" - If true, the request waits (long-polls) until the user responds to the prompt or more than 10 minutes pass. Defaults to false.

createPromptRequestExpiresIn :: !(Maybe Double)

"expires_in" - The number of seconds until this request can no longer be answered.

createPromptRequestMetadata :: !(Maybe PromptMetadata)

"metadata"

mkCreatePromptRequest

Arguments

:: Text
createPromptRequestUser: The user to send the approval request to. Can be either an email address or a phone number.

-> Text
createPromptRequestBody: The body of the approval request to show the user.

-> CreatePromptRequest

Construct a value of type CreatePromptRequest (by applying it's required fields, if any)

Error

data Error

Error

Constructors

Error

Fields

errorError :: !Text

Required "error" - A human readable API error message.

mkError

Arguments

:: Text
errorError: A human readable API error message.

-> Error

Construct a value of type Error (by applying it's required fields, if any)

Prompt

data Prompt

Prompt

Constructors

Prompt

Fields

promptId :: !Text

Required "id" - A unique id for this prompt.

promptSentAt :: !Double

Required "sent_at" - The unix timestamp when this prompt was sent.

promptIsExpired :: !Bool

Required "is_expired" - Whether the prompt can still be answered.

promptAnswer :: !(Maybe PromptAnswer)

"answer"

promptMetadata :: !(Maybe PromptMetadata)

"metadata"

mkPrompt

Arguments

:: Text
promptId: A unique id for this prompt.

-> Double
promptSentAt: The unix timestamp when this prompt was sent.

-> Bool
promptIsExpired: Whether the prompt can still be answered.

-> Prompt

Construct a value of type Prompt (by applying it's required fields, if any)

PromptAnswer

data PromptAnswer

PromptAnswer

Constructors

PromptAnswer

Fields

promptAnswerResult :: !Bool

Required "result" - The user's answer to whether or not they approve this prompt.

promptAnswerTime :: !Double

Required "time" - The unix timestamp when the user answered the prompt.

promptAnswerMetadata :: !(Maybe AnswerMetadata)

"metadata"

mkPromptAnswer

Arguments

:: Bool
promptAnswerResult: The user's answer to whether or not they approve this prompt.

-> Double
promptAnswerTime: The unix timestamp when the user answered the prompt.

-> PromptAnswer

Construct a value of type PromptAnswer (by applying it's required fields, if any)

PromptMetadata

data PromptMetadata

PromptMetadata

Constructors

PromptMetadata

Fields

promptMetadataLocation :: !(Maybe Text)

"location" - The physical location, like Oakland, CA, of the action.

promptMetadataTime :: !(Maybe Text)

"time" - The date/time of the action.

promptMetadataIpAddress :: !(Maybe Text)

"ip_address" - The IP address of the computer initiating the action.

promptMetadataBrowser :: !(Maybe Text)

"browser" - The web browser initiating the action, i.e. Chrome.

promptMetadataOperatingSystem :: !(Maybe Text)

"operating_system" - The operating system initiating the action, i.e. Mac OS X.

mkPromptMetadata :: PromptMetadata

Construct a value of type PromptMetadata (by applying it's required fields, if any)

PromptStatus

data PromptStatus

PromptStatus

Constructors

PromptStatus

Fields

promptStatusIsAnswered :: !Bool

Required "is_answered" - Whether the prompt has been answered or not.

promptStatusIsExpired :: !Bool

Required "is_expired" - Whether the prompt can still be answered.

mkPromptStatus

Arguments

:: Bool
promptStatusIsAnswered: Whether the prompt has been answered or not.

-> Bool
promptStatusIsExpired: Whether the prompt can still be answered.

-> PromptStatus

Construct a value of type PromptStatus (by applying it's required fields, if any)

Auth Methods

AuthBasicApiKey

data AuthBasicApiKey

Constructors

AuthBasicApiKey ByteString ByteString
username password